+++ visualizations/sunspot_contour_overlay.py
@@ -21,14 +21,14 @@ from matplotlib import cm
 directory = 'F:/Users/Brendan/Desktop/SolarProject'
 #date = '20141227'
 
-dates = ['20131118', '20140606', '20140818', '20141227', '20160414', '20160426', '20160520', '20160905', '20170329']
-contours = [110., 95., 75., 75., 87., 60., 60., 95., 83.]
-circ1_center_x = [520., 163.5, 99., 154., 540., 141., 327., 661., 512.]
-circ2_center_x = [520., 0, 99., 154., 0., 0., 327., 0., 0.]
-circ1_center_y = [290., 158., 81., 131., 433., 113., 245., 285., 350.]
-circ2_center_y = [290., 0, 81., 131., 0., 0., 245., 0., 0.]
-circ1_radi = [55., 50., 20., 24., 85., 26., 50., 51., 37.]
-circ2_radi = [73., 0., 39., 38., 0., 0., 73., 0., 0.]
+dates = ['20111210','20121018', '20131118', '20140112', '20140606', '20140818', '20140910', '20141227', '20150104','20160414', '20160426', '20160520', '20160905', '20170329']
+contours = [125., 95., 110., 67., 95., 75., 115., 75., 113., 87., 60., 60., 95., 83.]
+circ1_center_x = [0, 0, 520.,0,  163.5, 99.,0,  154., 0, 540., 141., 327., 661., 512.]
+circ2_center_x = [0, 0, 520.,0,  0, 99.,0,  154., 0, 0., 0., 327., 0., 0.]
+circ1_center_y = [0, 0, 290.,0,  158., 81.,0,  131., 0, 433., 113., 245., 285., 350.]
+circ2_center_y = [0, 0, 290.,0,  0, 81.,0,  131.,0,  0., 0., 245., 0., 0.]
+circ1_radi = [0, 0, 55.,0,  50., 20.,0,  24., 0, 85., 26., 50., 51., 37.]
+circ2_radi = [0, 0, 73.,0,  0., 39.,0,  38., 0, 0., 0., 73., 0., 0.]
 
 l = 8
 date = dates[l]    
@@ -49,7 +49,9 @@ titles = [r'Power Law Slope-Coefficient -- [$A$]', r'Power Law Index -- [$n$]', 
 names = ['slope_coeff', 'index', 'tail', 'gauss_amp', 'gauss_loc', 'gauss_wid', 'f_test', 'gauss_amp_scaled', 'p_value']
 
 heatmap = [h1,h2,h3,h4,h5]
+#heatmap = [h1,h3,h4,h5]
 wavelengths = [17100,1930,211,30400,160000]
+#wavelengths = [17100,211,30400,160000]
 #wavelengths = [171,193,211,304,1600]
 
 year = date[0:4]
@@ -98,6 +100,7 @@ for m in range(7):
 
 
 for c in range(len(heatmap)):
+#for c in range(1):
     h_map = heatmap[c]
     wavelength = wavelengths[c]
 
@@ -134,8 +137,8 @@ for c in range(len(heatmap)):
     v_min = np.percentile(visual,1)
     v_max = np.percentile(visual,99)
     
-    #for i in range(7):
-    for i in range(1,2):
+    for i in range(7):
+    #for i in range(1,2):
         
         
         if i == 6:
@@ -231,8 +234,50 @@ for c in range(len(heatmap)):
         #plt.tight_layout()
         #plt.savefig('%s/%s_%i_heatmap_%s.jpeg' % (path_name, date, wavelength, names[i]))
         #plt.savefig('%s/%s_%i_%s_same_%i.jpeg' % (path_name, date, wavelength, names[i], c))
-        #plt.savefig('%s/%s_%s_same_%i.jpeg' % (path_name, date, names[i], wavelength))
+        plt.savefig('%s/%s_%s_same_%i.jpeg' % (path_name, date, names[i], wavelength))
         #plt.savefig('%s/%s_%i_%s_same_%s.pdf' % (path_name, date, wavelength, names[i], seg[c]), format='pdf')
         #plt.close()
 
         #plt.title(r'%s %i$\AA$: Visual Average' % (dates[i],wavelength) + '\n Umbra [solid] + Penumbra [dashed]', y=1.01, fontsize=21)
+
+#fig = plt.figure(figsize=(13,9))
+fig = plt.figure(figsize=(fig_width,fig_height))
+v_min = np.percentile(visual,1)  # set heatmap vmin to 1% of data (could lower to 0.5% or 0.1%)
+v_max = np.percentile(visual,99)  # set heatmap vmax to 99% of data (could up to 99.5% or 99.9%)   
+ax = plt.gca()  # get current axis -- to set colorbar 
+plt.title(r'%s: %i $\AA$  [%s]' % (date_title, wavelength, titles[i]), y = 1.01, fontsize=25)
+#plt.title('%s -- Segment %i' % (titles[i], c), y = 1.01, fontsize=25)  # no date / wavelength
+#plt.title('%s -- Segment Hours: %s' % (titles[i], seg[c]), y = 1.01, fontsize=25)  # no date / wavelength
+circle = plt.Circle((circ1_center_x[l], circ1_center_y[l]), circ1_radi[l], linewidth=lw, color='white', fill=False, linestyle='dashed')
+circleB = plt.Circle((circ1_center_x[l], circ1_center_y[l]), circ1_radi[l], linewidth=lw+2.5, color='black', fill=False, linestyle='dashed')
+circle2 = plt.Circle((circ2_center_x[l], circ2_center_y[l]), circ2_radi[l], linewidth=lw, color='white', fill=False, linestyle='dashed')
+circle2B = plt.Circle((circ2_center_x[l], circ2_center_y[l]), circ2_radi[l], linewidth=lw+2.5, color='black', fill=False, linestyle='dashed')
+ax.add_artist(circleB)
+ax.add_artist(circle)
+ax.add_artist(circle2B)
+ax.add_artist(circle2)
+CS = plt.contour(X, Y, Z, levels=[contours[l]], linewidths=lw, colors='black', linestyles='solid')
+CS = plt.contour(X, Y, Z, levels=[contours[l]], linewidths=lw, colors='white', linestyles='dashed')
+#im = ax.imshow(np.flipud(h_map[i]), cmap = cmap, vmin=vmin[i], vmax=vmax[i])
+im = ax.imshow(visual, cmap = 'sdoaia1600', vmin=v_min, vmax=v_max)
+#im = ax.imshow(np.flipud(h_map2[i]), cmap = cmap, vmin=M2_low[i], vmax=M2_high[i])
+#plt.xlabel('X-position (i) [pixels]', fontsize=20, labelpad=10)
+#plt.ylabel('Y-position (j) [pixels]', fontsize=20, labelpad=10)
+plt.xticks(fontsize=17)
+plt.yticks(fontsize=17)
+divider = make_axes_locatable(ax)  # set colorbar to heatmap axis
+cax = divider.append_axes("right", size="3%", pad=0.07)
+if i == 0:
+    cbar = plt.colorbar(im,cax=cax, format='%0.2e')
+else:
+    cbar = plt.colorbar(im,cax=cax)
+#cbar.set_label('%s' % cbar_labels[i], size=20, labelpad=10)
+cbar.ax.tick_params(labelsize=17, pad=5) 
+#plt.tight_layout()
+#plt.savefig('%s/%s_%i_heatmap_%s.jpeg' % (path_name, date, wavelength, names[i]))
+#plt.savefig('%s/%s_%i_%s_same_%i.jpeg' % (path_name, date, wavelength, names[i], c))
+plt.savefig('%s/%s_visual_average_%i.jpeg' % (path_name, date, wavelength))
+#plt.savefig('%s/%s_%i_%s_same_%s.pdf' % (path_name, date, wavelength, names[i], seg[c]), format='pdf')
+#plt.close()
+
+#plt.title(r'%s %i$\AA$: Visual Average' % (dates[i],wavelength) + '\n Umbra [solid] + Penumbra [dashed]', y=1.01, fontsize=21)